Build the live tree by impling the preliminaries
It's rather weird how Rust deals with these kinds of things. In C++ we'd move some pointers around
and call it a day. I wish to move out of the structure, that sounds like a good solution. Like, moving
the entire struct out of something.
Unfortunately, this is not the right approach for this. I have a map containing the different keys and "chapter"s.

How do I efficiently extract elements?
Ideally, the map should be
Key -> Rc<Chapter>
This way, a Vec<Rc<Chapter>> can clone the Rc.
The main problem is internal mutability though. What a pain.